# frontend/openapi/admin-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "REST to MCP Admin API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/backend-apis/": {
      "get": {
        "tags": [
          "backend-apis"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "in": "query",
            "name": "pageSize",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "isActive",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "organizationId",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "organizationId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1
                          },
                          "slug": {
                            "type": "string",
                            "minLength": 1
                          },
                          "description": {
                            "type": "string",
                            "nullable": true
                          },
                          "defaultBaseUrl": {
                            "type": "string",
                            "format": "uri"
                          },
                          "authType": {
                            "type": "string",
                            "enum": [
                              "none",
                              "api_key",
                              "bearer",
                              "basic",
                              "oauth2"
                            ]
                          },
                          "authConfig": {
                            "type": "object",
                            "additionalProperties": {}
                          },
                          "apiKeyLocation": {
                            "type": "string",
                            "enum": [
                              "header",
                              "query"
                            ],
                            "nullable": true
                          },
                          "apiKeyName": {
                            "type": "string",
                            "nullable": true
                          },
                          "apiKeyValue": {
                            "type": "string",
                            "nullable": true
                          },
                          "hasApiKeyValue": {
                            "type": "boolean"
                          },
                          "apiKeyMaskedValue": {
                            "type": "string",
                            "nullable": true
                          },
                          "bearerToken": {
                            "type": "string",
                            "nullable": true
                          },
                          "hasBearerToken": {
                            "type": "boolean"
                          },
                          "basicUsername": {
                            "type": "string",
                            "nullable": true
                          },
                          "basicPassword": {
                            "type": "string",
                            "nullable": true
                          },
                          "hasBasicPassword": {
                            "type": "boolean"
                          },
                          "oauth2AccessToken": {
                            "type": "string",
                            "nullable": true
                          },
                          "hasOauth2AccessToken": {
                            "type": "boolean"
                          },
                          "defaultTimeoutMs": {
                            "type": "integer",
                            "exclusiveMinimum": true,
                            "minimum": 0
                          },
                          "retryPolicy": {
                            "type": "object",
                            "additionalProperties": {},
                            "nullable": true
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "organizationId",
                          "name",
                          "slug",
                          "defaultBaseUrl",
                          "authType",
                          "authConfig",
                          "defaultTimeoutMs",
                          "retryPolicy",
                          "isActive"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer",
                              "exclusiveMinimum": true,
                              "minimum": 0
                            },
                            "pageSize": {
                              "type": "integer",
                              "exclusiveMinimum": true,
                              "minimum": 0
                            },
                            "total": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "pageCount": {
                              "type": "integer",
                              "minimum": 0
                            }
                          },
                          "required": [
                            "page",
                            "pageSize",
                            "total",
                            "pageCount"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "pagination"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "backend-apis"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "organizationId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "defaultBaseUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "authType": {
                    "type": "string",
                    "enum": [
                      "none",
                      "api_key",
                      "basic",
                      "bearer",
                      "oauth2"
                    ],
                    "default": "none"
                  },
                  "authConfig": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "apiKeyLocation": {
                    "type": "string",
                    "enum": [
                      "header",
                      "query"
                    ]
                  },
                  "apiKeyName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "apiKeyValue": {
                    "type": "string",
                    "minLength": 1
                  },
                  "bearerToken": {
                    "type": "string",
                    "minLength": 1
                  },
                  "basicUsername": {
                    "type": "string",
                    "minLength": 1
                  },
                  "basicPassword": {
                    "type": "string",
                    "minLength": 1
                  },
                  "oauth2AccessToken": {
                    "type": "string",
                    "minLength": 1
                  },
                  "defaultTimeoutMs": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0,
                    "default": 30000
                  },
                  "retryPolicy": {
                    "type": "object",
                    "additionalProperties": {},
                    "default": {
                      "retries": 0
                    }
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "required": [
                  "organizationId",
                  "name",
                  "slug",
                  "defaultBaseUrl"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "organizationId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "slug": {
                          "type": "string",
                          "minLength": 1
                        },
                        "description": {
                          "type": "string",
                          "nullable": true
                        },
                        "defaultBaseUrl": {
                          "type": "string",
                          "format": "uri"
                        },
                        "authType": {
                          "type": "string",
                          "enum": [
                            "none",
                            "api_key",
                            "bearer",
                            "basic",
                            "oauth2"
                          ]
                        },
                        "authConfig": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "apiKeyLocation": {
                          "type": "string",
                          "enum": [
                            "header",
                            "query"
                          ],
                          "nullable": true
                        },
                        "apiKeyName": {
                          "type": "string",
                          "nullable": true
                        },
                        "apiKeyValue": {
                          "type": "string",
                          "nullable": true
                        },
                        "hasApiKeyValue": {
                          "type": "boolean"
                        },
                        "apiKeyMaskedValue": {
                          "type": "string",
                          "nullable": true
                        },
                        "bearerToken": {
                          "type": "string",
                          "nullable": true
                        },
                        "hasBearerToken": {
                          "type": "boolean"
                        },
                        "basicUsername": {
                          "type": "string",
                          "nullable": true
                        },
                        "basicPassword": {
                          "type": "string",
                          "nullable": true
                        },
                        "hasBasicPassword": {
                          "type": "boolean"
                        },
                        "oauth2AccessToken": {
                          "type": "string",
                          "nullable": true
                        },
                        "hasOauth2AccessToken": {
                          "type": "boolean"
                        },
                        "defaultTimeoutMs": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0
                        },
                        "retryPolicy": {
                          "type": "object",
                          "additionalProperties": {},
                          "nullable": true
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "organizationId",
                        "name",
                        "slug",
                        "defaultBaseUrl",
                        "authType",
                        "authConfig",
                        "defaultTimeoutMs",
                        "retryPolicy",
                        "isActive"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/backend-apis/{id}": {
      "patch": {
        "tags": [
          "backend-apis"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "defaultBaseUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "authType": {
                    "type": "string",
                    "enum": [
                      "none",
                      "api_key",
                      "basic",
                      "bearer",
                      "oauth2"
                    ],
                    "default": "none"
                  },
                  "authConfig": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "apiKeyLocation": {
                    "type": "string",
                    "enum": [
                      "header",
                      "query"
                    ]
                  },
                  "apiKeyName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "apiKeyValue": {
                    "type": "string",
                    "minLength": 1
                  },
                  "bearerToken": {
                    "type": "string",
                    "minLength": 1
                  },
                  "basicUsername": {
                    "type": "string",
                    "minLength": 1
                  },
                  "basicPassword": {
                    "type": "string",
                    "minLength": 1
                  },
                  "oauth2AccessToken": {
                    "type": "string",
                    "minLength": 1
                  },
                  "defaultTimeoutMs": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0,
                    "default": 30000
                  },
                  "retryPolicy": {
                    "type": "object",
                    "additionalProperties": {},
                    "default": {
                      "retries": 0
                    }
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "organizationId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "slug": {
                          "type": "string",
                          "minLength": 1
                        },
                        "description": {
                          "type": "string",
                          "nullable": true
                        },
                        "defaultBaseUrl": {
                          "type": "string",
                          "format": "uri"
                        },
                        "authType": {
                          "type": "string",
                          "enum": [
                            "none",
                            "api_key",
                            "bearer",
                            "basic",
                            "oauth2"
                          ]
                        },
                        "authConfig": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "apiKeyLocation": {
                          "type": "string",
                          "enum": [
                            "header",
                            "query"
                          ],
                          "nullable": true
                        },
                        "apiKeyName": {
                          "type": "string",
                          "nullable": true
                        },
                        "apiKeyValue": {
                          "type": "string",
                          "nullable": true
                        },
                        "hasApiKeyValue": {
                          "type": "boolean"
                        },
                        "apiKeyMaskedValue": {
                          "type": "string",
                          "nullable": true
                        },
                        "bearerToken": {
                          "type": "string",
                          "nullable": true
                        },
                        "hasBearerToken": {
                          "type": "boolean"
                        },
                        "basicUsername": {
                          "type": "string",
                          "nullable": true
                        },
                        "basicPassword": {
                          "type": "string",
                          "nullable": true
                        },
                        "hasBasicPassword": {
                          "type": "boolean"
                        },
                        "oauth2AccessToken": {
                          "type": "string",
                          "nullable": true
                        },
                        "hasOauth2AccessToken": {
                          "type": "boolean"
                        },
                        "defaultTimeoutMs": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0
                        },
                        "retryPolicy": {
                          "type": "object",
                          "additionalProperties": {},
                          "nullable": true
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "organizationId",
                        "name",
                        "slug",
                        "defaultBaseUrl",
                        "authType",
                        "authConfig",
                        "defaultTimeoutMs",
                        "retryPolicy",
                        "isActive"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "backend-apis"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/admin/v1"
    }
  ]
}